MCS-3 - create dumb components for conditional rendering
