+++ src/MarvelHeader.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import './App.css';
+
+function MarvelHeader() {
+    return (
+        <header className="App-header">
+            <div className="App-logo-header">
+                <p>Marvel Comics Store.</p>
+            </div>
+        </header>
+    );
+}
+
+export default MarvelHeader;
